# docs/integration/openapi-iteration3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "教育服务系统 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/crm/leads": {
      "post": {
        "tags": [
          "企业助手"
        ],
        "summary": "新增意向客户",
        "description": "新增意向客户。\n\n**业务规则：**\n- customer_name 必填\n- 新客户默认状态 new\n- owner_employee_id 默认为当前登录用户（当前版本: 从请求体获取）\n- 插入前校验 owner_employee_id 对应的 sys_user 存在且 user_type='employee'",
        "operationId": "create_lead_api_v1_crm_leads_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "企业助手"
        ],
        "summary": "查询意向客户（支持条件搜索+分页）",
        "description": "查询意向客户列表。\n\n**权限规则：**\n- 员工：只能看到 owner_employee_id = current_user.id 的客户\n- 经理：可查看本部门员工的客户\n- 管理员：可查看全部客户\n（当前版本暂未实现完整权限过滤，通过 owner_employee_id 参数自行筛选）\n\n**Dify 白名单：** 此接口在白名单中，Dify 可通过 Service Token 调用。",
        "operationId": "list_leads_api_v1_crm_leads_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态筛选: new/contacting/qualified/signed/lost",
              "title": "Status"
            },
            "description": "按状态筛选: new/contacting/qualified/signed/lost"
          },
          {
            "name": "owner_employee_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按负责人筛选",
              "title": "Owner Employee Id"
            },
            "description": "按负责人筛选"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按姓名/联系方式模糊搜索",
              "title": "Keyword"
            },
            "description": "按姓名/联系方式模糊搜索"
          },
          {
            "name": "create_time_start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "创建时间起始",
              "title": "Create Time Start"
            },
            "description": "创建时间起始"
          },
          {
            "name": "create_time_end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "创建时间截止",
              "title": "Create Time End"
            },
            "description": "创建时间截止"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crm/leads/{lead_id}": {
      "get": {
        "tags": [
          "企业助手"
        ],
        "summary": "查询单个客户详情",
        "operationId": "get_lead_api_v1_crm_leads__lead_id__get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "企业助手"
        ],
        "summary": "更新客户信息",
        "operationId": "update_lead_api_v1_crm_leads__lead_id__put",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crm/leads/{lead_id}/status": {
      "put": {
        "tags": [
          "企业助手"
        ],
        "summary": "更新客户状态（含业务规则校验）",
        "description": "更新客户状态。\n\n**状态机约束：**\n```\nnew → contacting → qualified → signed（终态）\n  │       │           │\n  └───────┴───────────┴──────────→ lost（终态，必须填 lost_reason）\n```\n- signed 和 lost 为终态，不能回退\n- 变更为 lost 时 lost_reason 必填\n- 使用条件 UPDATE 防并发状态覆盖",
        "operationId": "update_lead_status_api_v1_crm_leads__lead_id__status_put",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crm/leads/{lead_id}/follow-ups": {
      "post": {
        "tags": [
          "企业助手"
        ],
        "summary": "新增跟进记录",
        "description": "新增跟进记录。\n\n**业务规则：**\n- lead_id 对应的 crm_lead 记录必须存在（应用层逻辑外键校验）\n- content 为必填\n- 同步更新 crm_lead.last_contact_time 和 update_time",
        "operationId": "create_follow_up_api_v1_crm_leads__lead_id__follow_ups_post",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "企业助手"
        ],
        "summary": "查询跟进历史",
        "operationId": "list_follow_ups_api_v1_crm_leads__lead_id__follow_ups_get",
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assistant/chat": {
      "post": {
        "tags": [
          "智能助手",
          "智能助手"
        ],
        "summary": "发送消息",
        "description": "与智能助手对话。\n\n**支持的意图类型：**\n- 查询数据（NL2SQL）：如\"有多少个联系中的客户\"、\"最近一周新增了几个客户\"\n- 执行操作（NL2API）：如\"帮我录入新客户王五\"、\"更新客户3号的状态为联系中\"\n- 闲聊：其他问题\n\n**首次对话：** 不传 session_id，后端自动创建新会话\n**多轮对话：** 传入之前返回的 session_id，保持上下文",
        "operationId": "chat_api_v1_assistant_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/assistant/sessions": {
      "get": {
        "tags": [
          "智能助手",
          "智能助手"
        ],
        "summary": "会话列表",
        "description": "查询当前员工的智能助手会话列表。",
        "operationId": "list_sessions_api_v1_assistant_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/assistant/sessions/{session_id}/messages": {
      "get": {
        "tags": [
          "智能助手",
          "智能助手"
        ],
        "summary": "会话历史",
        "description": "查询某个会话的消息历史记录。",
        "operationId": "list_messages_api_v1_assistant_sessions__session_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadCreate": {
        "properties": {
          "customer_name": {
            "type": "string",
            "maxLength": 64,
            "title": "Customer Name",
            "examples": [
              "李四"
            ]
          },
          "contact_info": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Info",
            "examples": [
              "139xxxx0002"
            ]
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "examples": [
              "M"
            ]
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "examples": [
              25
            ]
          },
          "education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education Level",
            "examples": [
              "本科"
            ]
          },
          "intended_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Country",
            "examples": [
              "英国,美国"
            ]
          },
          "intended_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Major",
            "examples": [
              "计算机科学"
            ]
          },
          "source_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Channel",
            "examples": [
              "线上"
            ]
          },
          "background_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Info",
            "examples": [
              "对UCL计算机硕士感兴趣，GPA 3.5"
            ]
          },
          "customer_profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Profile Id",
            "examples": [
              1
            ]
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "examples": [
              "高意向客户，需尽快跟进"
            ]
          },
          "owner_employee_id": {
            "type": "integer",
            "title": "Owner Employee Id",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "customer_name",
          "owner_employee_id"
        ],
        "title": "LeadCreate",
        "description": "新增意向客户"
      },
      "LeadStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "contacting",
              "qualified",
              "signed",
              "lost"
            ],
            "title": "Status",
            "description": "客户状态：new(新)/contacting(跟进中)/qualified(已确认)/signed(已签约)/lost(已流失)",
            "examples": [
              "contacting"
            ]
          },
          "lost_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Lost Reason",
            "description": "仅 lost 时填写",
            "examples": [
              "价格太高"
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LeadStatusUpdate",
        "description": "更新客户状态"
      },
      "LeadUpdate": {
        "properties": {
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "contact_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Info"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education Level"
          },
          "intended_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Country"
          },
          "intended_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Major"
          },
          "source_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Channel"
          },
          "background_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Info"
          },
          "customer_profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Profile Id"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "owner_employee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Employee Id"
          }
        },
        "type": "object",
        "title": "LeadUpdate",
        "description": "更新意向客户信息（部分更新：所有字段 Optional，只更新传了的字段）"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "DifyServiceToken": {
        "type": "http",
        "description": "Dify HTTP 节点调用 FastAPI 白名单接口的服务令牌",
        "scheme": "bearer"
      }
    }
  }
}